feat(type-designer): rebase work on main
