--- packages/plugins/type-designer/src/components/elements-type-container/column-type.d.ts
@@ -1,7 +1,0 @@
-import type { DataTypeTemplates } from './data-type-templates'
-
-export type Column = {
-	name: string
-	visible: boolean
-	items: DataTypeTemplates[]
-}
