+++ packages/plugins/type-designer/src/stores/columns-type.store.d.ts
@@ -23,3 +23,9 @@ export type DataTypeTemplates =
 	| BayType
 	| IEDType
 	| LDeviceType
+
+export type Column = {
+	name: string
+	visible: boolean
+	items: DataTypeTemplates[]
+}
