+++ packages/plugins/type-designer/src/stores/index.ts
@@ -5,3 +5,4 @@ export * from './events.store'
 // CUSTOM
 
 export * from './data-types-templates.store'
+export * from './columns.store'
